Added support for RFC5652 types of MS countersignatures (#16) (#17)
* Added support for RFC5652 types of MS countersignatures (#16)

It seems that some (maybe newer) MS countersignatures are not PKCS7 per
RFC2315 but rather CMS structures defined by RFC5652. Unfortunately,
PKCS7_* family of OpenSSL functions is not able to handle it, but there
are CMS_* functions which are. They however do not provide same set of
functions to do the same things as with PKCS7 structures.

This PR adds possibility to fall back to CMS if PKCS7 fails. Some
functions had to be simulated and therefore might not be *that* accurate
but just from testing on a few files, it seems to validate them
correctly. But it might need a tuning in the future if we test it on a
more extensive set of samples.

* Removed debug prints

* CMake adjustments in hopes to fix Windows build

* Debugging GitHub Workflow issue

* Revert "CMake adjustments in hopes to fix Windows build"

This reverts commit b6b9bf3cc2d89525b7643f87eafe4923a0631d19.

* Revert "Debugging GitHub Workflow issue"

This reverts commit f644da401e8cd9f0124065df9c454436abeb4a0e.

* Include opensslv.h for OPENSSL_VERSION_NUMBER

* Format changes using clang-formatter

* Remove and move openssl to correct folder on Windows build

* Add python scripts for test dev purposes

* Add program name to authenticode dumper

* Add tests for new MS countersignature feature

* Update CI to use expected OpenSSL versions

---------

Co-authored-by: Karel Hájek <karel.hajek@gendigital.com>